# Micro optimize extend